requirements.txt
flask 
firebase-admin
numpy
pandas
matplotlib
scikit-learn
requests
python-dotenv
keras
tensorflow